Create new viewer
