+++ package.json
@@ -19,6 +19,7 @@
   },
   "description": "",
   "dependencies": {
+    "@thatopen/components": "~2.2.0",
     "@thatopen/ui": "~2.2.0",
     "bim-fragment": "1.0.22",
     "firebase": "10.5.2",
@@ -26,8 +27,8 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-router-dom": "6.18.0",
-    "three": "0.152.2",
+    "three": "0.160.1",
     "uuid": "^9.0.1",
-    "web-ifc": "0.0.44"
+    "web-ifc": "0.0.57"
   }
 }
